SVN: merge 3.0->trunk(3.1) (rev. 9461:10362,10386:10393)
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@10616 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/database/dbf/DBFFieldMetadata.java
@@ -27,7 +27,7 @@ public class DBFFieldMetadata{
 	public String name;
 	public char type;
 	public int offset;
-	public byte length;
-	public byte decPlaces;
+	public short length;
+	public short decPlaces;
 	
 }
